refactor: Move periodic aggregator to its own sub-module.
